requirements.txt
asgiref==3.6.0
boto3==1.26.102
botocore==1.29.102
dj-database-url==1.3.0
Django==4.1.7
django-cors-headers==3.14.0
django-dotenv==1.4.2
django-filter==23.1
django-storages==1.13.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
jmespath==1.0.1
psycopg2==2.9.5
PyJWT==2.6.0
python-dateutil==2.8.2
pytz==2023.3
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.5.0
urllib3==1.26.15
whitenoise==6.4.0
